+++ backend/core/src/seeds/listings/listing-default-seed.ts
@@ -21,6 +21,8 @@ import {
   getLiveWorkPreference,
 } from "./shared"
 import { ApplicationMethod } from "../../application-methods/entities/application-method.entity"
+import { Jurisdiction } from "../../jurisdictions/entities/jurisdiction.entity"
+import { CountyCode } from "../../shared/types/county-code"
 
 export class ListingDefaultSeed {
   constructor(
@@ -37,7 +39,9 @@ export class ListingDefaultSeed {
     @InjectRepository(Unit) protected readonly unitsRepository: Repository<Unit>,
     @InjectRepository(User) protected readonly userRepository: Repository<User>,
     @InjectRepository(ApplicationMethod)
-    protected readonly applicationMethodRepository: Repository<ApplicationMethod>
+    protected readonly applicationMethodRepository: Repository<ApplicationMethod>,
+    @InjectRepository(Jurisdiction)
+    protected readonly jurisdictionRepository: Repository<Jurisdiction>
   ) {}
 
   async seed() {
@@ -46,7 +50,13 @@ export class ListingDefaultSeed {
     )
     const unitTypeOneBdrm = await this.unitTypeRepository.findOneOrFail({ name: "oneBdrm" })
     const unitTypeTwoBdrm = await this.unitTypeRepository.findOneOrFail({ name: "twoBdrm" })
-    const amiChart = await this.amiChartRepository.save(getDefaultAmiChart())
+    const alamedaJurisdiction = await this.jurisdictionRepository.findOneOrFail({
+      name: CountyCode.alameda,
+    })
+    const amiChart = await this.amiChartRepository.save({
+      ...getDefaultAmiChart(),
+      jurisdiction: alamedaJurisdiction,
+    })
 
     const property = await this.propertyRepository.save({
       ...getDefaultProperty(),