+++ backend/core/src/seeds/listings/shared.ts
@@ -28,7 +28,7 @@ export function getDefaultAmiChart() {
   return JSON.parse(JSON.stringify(defaultAmiChart))
 }
 
-export const defaultAmiChart: AmiChartCreateDto = {
+export const defaultAmiChart: Omit<AmiChartCreateDto, "jurisdiction"> = {
   name: "AlamedaCountyTCAC2021",
   items: [
     {